# docs/swagger.yaml
definitions:
  types.BucketInfo:
    properties:
      creation_date:
        description: 桶的创建时间
        type: string
      location:
        description: Bucket datacenter
        type: string
      name:
        description: 桶的名称
        type: string
      owner:
        description: 桶的所有者
        type: string
      region:
        description: Bucket region
        type: string
      storage_class:
        description: Bucket storage class
        type: string
    type: object
  types.ObjectInfo:
    properties:
      etag:
        description: 对象的 ETag
        type: string
      hear:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      last_modified:
        description: 对象最后修改时间
        type: string
      name:
        description: 对象的名称
        type: string
      size:
        description: 对象的⼤⼩（字节）
        type: integer
      storage_class:
        type: string
    type: object
  types.ObjectMetadata:
    properties:
      bucket_name:
        description: 对象所属的桶名称
        type: string
      content_type:
        description: 对象的内容类型
        type: string
      e_tag:
        description: 对象的 ETag （通常是内容的 MD5 哈希）
        type: string
      is_latest:
        description: 是否是最新版本
        type: boolean
      last_modified:
        description: 对象最后修改时间
        type: string
      object_name:
        description: 对象的名称
        type: string
      size:
        description: 对象的⼤⼩（字节）
        type: integer
      storage_nodes:
        description: 存储该对象的节点列表
        items:
          type: string
        type: array
      version_id:
        description: 对象的版本 ID （如果启⽤了版本控制）
        type: string
    type: object
info:
  contact: {}
paths:
  /metadata/:name:
    delete:
      consumes:
      - application/json
      description: 根据 name 删除Bucket
      parameters:
      - description: Bucket名字
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: 删除Bucket
      tags:
      - metadata
    post:
      consumes:
      - application/json
      description: 根据 name 创建Bucket
      parameters:
      - description: Bucket名字
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: 创建Bucket
      tags:
      - metadata
  /metadata/bucket/list:
    get:
      consumes:
      - application/json
      description: 根据 prefix 和 max_keys 查询Bucket
      parameters:
      - description: Prefix
        in: query
        name: prefix
        type: string
      - description: Maximum number of keys to return
        in: query
        name: max_keys
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.BucketInfo'
            type: array
        "400":
          description: Bad Request
      summary: 获取Bucket列表
      tags:
      - metadata
  /metadata/object:
    get:
      consumes:
      - application/json
      description: 根据 bucket_name 和 object_name 查询对象元数据信息
      parameters:
      - in: query
        name: bucket_name
        type: string
      - in: query
        name: file_path
        type: string
      - in: query
        name: object_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ObjectMetadata'
        "400":
          description: Bad Request
      summary: 获取对象元数据信息
      tags:
      - metadata
  /metadata/object/list:
    get:
      consumes:
      - application/json
      description: 根据 bucket_name、prefix 和 max_keys 查询对象元数据
      parameters:
      - description: Bucket Name
        in: query
        name: bucket_name
        required: true
        type: string
      - description: Prefix
        in: query
        name: prefix
        type: string
      - description: Maximum number of keys to return
        in: query
        name: max_keys
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.ObjectInfo'
            type: array
        "400":
          description: Bad Request
      summary: 获取对象元数据列表
      tags:
      - metadata
  /storage/delete:
    delete:
      consumes:
      - application/json
      description: 根据 bucket_name 和object_name 删除文件
      parameters:
      - in: query
        name: bucket_name
        type: string
      - in: query
        name: file_path
        type: string
      - in: query
        name: object_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: 删除文件
      tags:
      - storage
  /storage/object:
    get:
      consumes:
      - application/json
      description: 根据 bucket_name 和 object_name 查询文件信息
      parameters:
      - in: query
        name: bucket_name
        type: string
      - in: query
        name: file_path
        type: string
      - in: query
        name: object_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回上传的文件信息
          schema:
            type: object
        "400":
          description: Bad Request
      summary: 获取文件信息
      tags:
      - storage
  /storage/upload:
    post:
      consumes:
      - multipart/form-data
      description: 根据 bucket_name 和 object_name 上传文件
      parameters:
      - description: Bucket Name
        in: formData
        name: bucket_name
        required: true
        type: string
      - description: Object Name
        in: formData
        name: object_name
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回上传的文件ETag
          schema:
            type: string
        "400":
          description: 错误响应
      summary: 上传文件
      tags:
      - storage
swagger: "2.0"